requirements.txt
python==3.7.13
pytorch==1.12.0
numpy==1.21.5
scikit-learn==0.22.2.post1
scipy==1.7.3